+++ infra/gorm/repository/user_repository_test.go
@@ -4,9 +4,9 @@ import (
 	"testing"
 
 	"github.com/gghcode/apas-todo-apiserver/config"
-	"github.com/gghcode/apas-todo-apiserver/db"
 	"github.com/gghcode/apas-todo-apiserver/domain/entity"
 	"github.com/gghcode/apas-todo-apiserver/domain/usecase/user"
+	"github.com/gghcode/apas-todo-apiserver/infra/gorm"
 	"github.com/gghcode/apas-todo-apiserver/infra/gorm/model"
 	"github.com/gghcode/apas-todo-apiserver/infra/gorm/repository"
 	"github.com/gghcode/apas-todo-apiserver/internal/testutil"
@@ -34,7 +34,7 @@ func (suite *userRepositoryIntegrationTestSuite) SetupTest() {
 	cfg, err := config.FromEnvs()
 	suite.NoError(err)
 
-	dbConn, _, err := db.NewPostgresConn(cfg)
+	dbConn, _, err := gorm.NewPostgresConn(cfg)
 	suite.NoError(err)
 
 	suite.dbCleanup = testutil.DbCleanupFunc(dbConn.DB())
